+++ utils/produce_html_plotpage.cc
@@ -15,13 +15,12 @@
 #include <string>
 #include <vector>
 #include <assert.h>
-
 #include <complex>
-#include <Log_writer_cout.h>
-#include <gnuplot_i.h>
-#include <utils.h>
 
-#include <Control_parameters.h>
+#include "log_writer_cout.h"
+#include "gnuplot_i.h"
+#include "utils.h"
+#include "control_parameters.h"
 
 Control_parameters ConPrms;
 
@@ -32,8 +31,8 @@ public:
   double frequency;
   char sideband;
   std::vector<std::string> autos, crosses;
-  std::vector<DOUBLE> snr_crosses; // Signal to noise ratio for the crosses
-  std::vector<DOUBLE> offset; // Offset of the maximum value of a cross with the midpoint
+  std::vector<FLOAT> snr_crosses; // Signal to noise ratio for the crosses
+  std::vector<FLOAT> offset; // Offset of the maximum value of a cross with the midpoint
 
   void set_size_crosses(int size_cross) {
     crosses.resize(size_cross);
@@ -68,13 +67,13 @@ private:
                            const Control_parameters &ConPrms);
   void plot(char *filename, int nPts, char *title);
 
-  DOUBLE signal_to_noise_ratio(std::vector< std::complex<DOUBLE> > &data);
-  DOUBLE max_value_offset(std::vector< std::complex<DOUBLE> > &data);
+  FLOAT signal_to_noise_ratio(std::vector< std::complex<FLOAT> > &data);
+  FLOAT max_value_offset(std::vector< std::complex<FLOAT> > &data);
 
 private:
   int nLags;
-  std::vector< std::complex<DOUBLE> > in, out;
-  std::vector< DOUBLE > magnitude;
+  std::vector< std::complex<FLOAT> > in, out;
+  std::vector< FLOAT > magnitude;
   FFTW_PLAN visibilities2lags; 
 };
 
@@ -285,7 +284,7 @@ Plot_generator::generate_auto_plots(std::ifstream &infile,
                                     const Control_parameters &ConPrms) {
 
   for (int station=stations_start; station<stations_end; station++) {
-    infile.read((char *)&in[0], 2*in.size()*sizeof(DOUBLE));
+    infile.read((char *)&in[0], 2*in.size()*sizeof(FLOAT));
 
     for  (int lag=0; lag<nLags; lag++) {
       magnitude[lag] = abs(in[lag]);
@@ -313,7 +312,7 @@ Plot_generator::generate_cross_plot(std::ifstream &infile,
                                     const Control_parameters &ConPrms) {
   int nStations = ConPrms.number_stations();
 
-  infile.read((char *)&in[0], 2*in.size()*sizeof(DOUBLE));
+  infile.read((char *)&in[0], 2*in.size()*sizeof(FLOAT));
   FFTW_EXECUTE(visibilities2lags);
 
   for  (int lag=0; lag<nLags; lag++) {
@@ -361,15 +360,15 @@ Plot_generator::plot(char *filename, int nPts, char *title) {
   gnuplot_close(g);
 }
 
-DOUBLE
-Plot_generator::max_value_offset(std::vector< std::complex<DOUBLE> > &data)
+FLOAT
+Plot_generator::max_value_offset(std::vector< std::complex<FLOAT> > &data)
 {
   int index_max = 0;
   for (size_t i=1; i<data.size(); i++) {
     if (norm(data[i]) > norm(data[index_max])) index_max = i;
   }
 
-  DOUBLE maxval = 0.0;
+  FLOAT maxval = 0.0;
   int maxval_loc = 0;
   for  (int lag=0; lag<data.size(); lag++) {
     magnitude[lag] = abs(data[(lag+data.size()/2)%data.size()])/data.size();
@@ -381,8 +380,8 @@ Plot_generator::max_value_offset(std::vector< std::complex<DOUBLE> > &data)
   
   return maxval_loc;
 }
-DOUBLE 
-Plot_generator::signal_to_noise_ratio(std::vector< std::complex<DOUBLE> > &data)
+FLOAT 
+Plot_generator::signal_to_noise_ratio(std::vector< std::complex<FLOAT> > &data)
 {
   int index_max = 0;
   for (size_t i=1; i<data.size(); i++) {
@@ -390,7 +389,7 @@ Plot_generator::signal_to_noise_ratio(std::vector< std::complex<DOUBLE> > &data)
   }
 
   //return noise rms in array, skip 10 % around maximum
-  std::complex<DOUBLE> mean(0,0);
+  std::complex<FLOAT> mean(0,0);
   int ll=index_max - data.size()/20;//5% of range to left
   int ul=index_max + data.size()/20;//5% of range to right
   int n2avg=0;
@@ -406,7 +405,7 @@ Plot_generator::signal_to_noise_ratio(std::vector< std::complex<DOUBLE> > &data)
 
   mean /= n2avg;
 
-  DOUBLE sum = 0;
+  FLOAT sum = 0;
   for (size_t i=0 ; i< data.size() ; i++){
     if (((int)i < ll) || ((int)i > ul)) {
       sum += norm(data[i]-mean);
